+++ mobile/lib/pages/app_log.page.dart
