+++ mobile/lib/pages/app_log_detail.page.dart
