+++ mobile/lib/pages/search/all_places.page.dart
