+++ mobile/lib/pages/search/person_result.page.dart
