--- mobile/lib/pages/settings_sub_page.dart
@@ -1,22 +1,0 @@
-import 'package:auto_route/auto_route.dart';
-import 'package:easy_localization/easy_localization.dart';
-import 'package:flutter/material.dart';
-import 'package:immich_mobile/pages/settings_page.dart';
-
-@RoutePage()
-class SettingsSubPage extends StatelessWidget {
-  const SettingsSubPage(this.section, {super.key});
-
-  final SettingSection section;
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        centerTitle: false,
-        title: Text(section.title).tr(),
-      ),
-      body: section.widget,
-    );
-  }
-}
