+++ mobile/lib/pages/sharing/partner/partner_detail.page.dart
