+++ mobile/lib/routing/router.dart
@@ -3,15 +3,15 @@ import 'package:flutter/material.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:immich_mobile/models/memories/memory.model.dart';
 import 'package:immich_mobile/models/search/search_filter.model.dart';
-import 'package:immich_mobile/pages/activities_page.dart';
+import 'package:immich_mobile/pages/activities.page.dart';
 import 'package:immich_mobile/models/albums/asset_selection_page_result.model.dart';
-import 'package:immich_mobile/pages/album_options_part.dart';
-import 'package:immich_mobile/pages/album_viewer_page.dart';
-import 'package:immich_mobile/pages/album_asset_selection_page.dart';
-import 'package:immich_mobile/pages/create_album_page.dart';
+import 'package:immich_mobile/pages/album_options.page.dart';
+import 'package:immich_mobile/pages/album_viewer.page.dart';
+import 'package:immich_mobile/pages/album_asset_selection.page.dart';
+import 'package:immich_mobile/pages/create_album.page.dart';
 import 'package:immich_mobile/models/shared_link/shared_link.model.dart';
-import 'package:immich_mobile/pages/album_additional_shared_user_selection_age.dart';
-import 'package:immich_mobile/pages/album_shared_user_selection_page.dart';
+import 'package:immich_mobile/pages/album_additional_shared_user_selection.page.dart';
+import 'package:immich_mobile/pages/album_shared_user_selection.page.dart';
 import 'package:immich_mobile/providers/gallery_permission.provider.dart';
 import 'package:immich_mobile/routing/auth_guard.dart';
 import 'package:immich_mobile/routing/custom_transition_builders.dart';
@@ -22,43 +22,42 @@ import 'package:immich_mobile/entities/album.entity.dart';
 import 'package:immich_mobile/entities/logger_message.entity.dart';
 import 'package:immich_mobile/entities/user.entity.dart';
 import 'package:immich_mobile/providers/api.provider.dart';
-import 'package:immich_mobile/pages/app_log_detail_page.dart';
-import 'package:immich_mobile/pages/app_log_page.dart';
-import 'package:immich_mobile/pages/backup/album_preview_page.dart';
-import 'package:immich_mobile/pages/backup/backup_album_selection_page.dart';
-import 'package:immich_mobile/pages/backup/backup_controller_page.dart';
-import 'package:immich_mobile/pages/backup/backup_options_page.dart';
-import 'package:immich_mobile/pages/backup/failed_backup_status_page.dart';
-import 'package:immich_mobile/pages/gallery_viewer.dart';
-import 'package:immich_mobile/pages/library/archive_page.dart';
-import 'package:immich_mobile/pages/library/favorites_page.dart';
-import 'package:immich_mobile/pages/library/library_page.dart';
-import 'package:immich_mobile/pages/library/trash_page.dart';
-import 'package:immich_mobile/pages/login/change_password_page.dart';
-import 'package:immich_mobile/pages/login/login_page.dart';
-import 'package:immich_mobile/pages/onboarding/permission_onboarding_page.dart';
-import 'package:immich_mobile/pages/photos/memory_page.dart';
-import 'package:immich_mobile/pages/photos/photos_page.dart';
-import 'package:immich_mobile/pages/search/all_motion_videos_page.dart';
-import 'package:immich_mobile/pages/search/all_people_page.dart';
-import 'package:immich_mobile/pages/search/all_places_page.dart';
-import 'package:immich_mobile/pages/search/all_videos_page.dart';
-import 'package:immich_mobile/pages/search/map/map_location_picker_page.dart';
-import 'package:immich_mobile/pages/search/map/map_page.dart';
-import 'package:immich_mobile/pages/search/person_result_page.dart';
-import 'package:immich_mobile/pages/search/recently_added_page.dart';
-import 'package:immich_mobile/pages/search/search_input_page.dart';
-import 'package:immich_mobile/pages/search/search_page.dart';
-import 'package:immich_mobile/pages/settings_page.dart';
-import 'package:immich_mobile/pages/settings_sub_page.dart';
-import 'package:immich_mobile/pages/sharing/partner/partner_detail_page.dart';
-import 'package:immich_mobile/pages/sharing/partner/partner_page.dart';
-import 'package:immich_mobile/pages/sharing/shared_link/shared_link_edit_page.dart';
-import 'package:immich_mobile/pages/sharing/shared_link/shared_link_page.dart';
-import 'package:immich_mobile/pages/sharing/sharing_page.dart';
-import 'package:immich_mobile/pages/splash_screen.dart';
-import 'package:immich_mobile/pages/tab_controller_page.dart';
-import 'package:immich_mobile/pages/video_viewer_page.dart';
+import 'package:immich_mobile/pages/app_log_detail.page.dart';
+import 'package:immich_mobile/pages/app_log.page.dart';
+import 'package:immich_mobile/pages/backup/album_preview.page.dart';
+import 'package:immich_mobile/pages/backup/backup_album_selection.page.dart';
+import 'package:immich_mobile/pages/backup/backup_controller.page.dart';
+import 'package:immich_mobile/pages/backup/backup_options.page.dart';
+import 'package:immich_mobile/pages/backup/failed_backup_status.page.dart';
+import 'package:immich_mobile/pages/gallery_viewer.page.dart';
+import 'package:immich_mobile/pages/library/archive.page.dart';
+import 'package:immich_mobile/pages/library/favorite.page.dart';
+import 'package:immich_mobile/pages/library/library.page.dart';
+import 'package:immich_mobile/pages/library/trash.page.dart';
+import 'package:immich_mobile/pages/login/change_password.page.dart';
+import 'package:immich_mobile/pages/login/login.page.dart';
+import 'package:immich_mobile/pages/onboarding/permission_onboarding.page.dart';
+import 'package:immich_mobile/pages/photos/memory.page.dart';
+import 'package:immich_mobile/pages/photos/photos.page.dart';
+import 'package:immich_mobile/pages/search/all_motion_videos.page.dart';
+import 'package:immich_mobile/pages/search/all_people.page.dart';
+import 'package:immich_mobile/pages/search/all_places.page.dart';
+import 'package:immich_mobile/pages/search/all_videos.page.dart';
+import 'package:immich_mobile/pages/search/map/map_location_picker.page.dart';
+import 'package:immich_mobile/pages/search/map/map.page.dart';
+import 'package:immich_mobile/pages/search/person_result.page.dart';
+import 'package:immich_mobile/pages/search/recently_added.page.dart';
+import 'package:immich_mobile/pages/search/search_input.page.dart';
+import 'package:immich_mobile/pages/search/search.page.dart';
+import 'package:immich_mobile/pages/settings.page.dart';
+import 'package:immich_mobile/pages/sharing/partner/partner_detail.page.dart';
+import 'package:immich_mobile/pages/sharing/partner/partner.page.dart';
+import 'package:immich_mobile/pages/sharing/shared_link/shared_link_edit.page.dart';
+import 'package:immich_mobile/pages/sharing/shared_link/shared_link.page.dart';
+import 'package:immich_mobile/pages/sharing/sharing.page.dart';
+import 'package:immich_mobile/pages/splash_screen.page.dart';
+import 'package:immich_mobile/pages/tab_controller.page.dart';
+import 'package:immich_mobile/pages/video_viewer.page.dart';
 import 'package:immich_mobile/services/api.service.dart';
 import 'package:isar/isar.dart';
 import 'package:maplibre_gl/maplibre_gl.dart';
