+++ mobile/lib/utils/immich_loading_overlay.dart
